Remove dashes from Python filenames
One cannot easily import a Python module if the name has dashes in it.
That tells me that Python would rather I not use those characters in
filenames.

Also changed:
FindCRCs: has improved output, especially in the batch script.
PhotoBackup: added logging and improved protection against deleting
EVERYTHING.

+++ PhotoBackup.py
@@ -27,34 +27,51 @@ filespec = "*"  # which files get archived?
 import glob, re, os, sys
 from crc32 import crc32
 from EXIF_Dating import GetExifDate, GetFileDate, GetExifDimensions
-from PyBackup import ReadCrcs, recursive_mkdir
+from PyBackup import logging, ReadCrcs, recursive_mkdir, display_update
+
+log = None  # A logging class, hopefully
+
+def MyLog(str):
+    """ Log a msg """
+    global log
+    if log != None:
+        log.msg(str)
 
 def remove_duplicate(src, dest):
     """ Remove a source file that does not need to be backed up """
     global archived, removed, not_archived
     try:
+        # Don't try to remove a file when the new pathname is identical
+        if src.lower() == dest.lower():
+            return
         os.remove(src)
         print("{} removed as a duplicate of {}".format(src, dest))
         removed += 1
     except:
-        print("exception: {} cannot be removed ".format(src))
+        MyLog("exception: {} cannot be removed ".format(src))
         not_archived += 1
 
 def replace_destination(src, dest, reason):
     """ Replace a destination file with a source file for a specified reason """
     global archived, not_archived
     try:
+        # Don't try to replqace a file when the new pathname is identical
+        if src.lower() == dest.lower():
+            return
         os.remove(dest)
     except:
-        print("{}: cannot remove ".format(dest))
+        MyLog("{}: cannot remove".format(dest))
         not_archived += 1
         return
     try:
+        # Don't try to rename a file when the new pathname is identical
+        if src.lower() == dest.lower():
+            return
         os.rename(src, dest)
         archived += 1
-        print("{} replaced: {}".format(dest, reason))
+        print("{} replaced {}".format(dest, reason))
     except:
-        print("{} cannot rename to {}".format(src, dest))
+        MyLog("{} cannot rename to {}".format(src, dest))
         not_archived += 1
 
 def archive(pn, recursive):
@@ -67,7 +84,6 @@ def archive(pn, recursive):
     rootp, ext = os.path.splitext(pn)
     rootp, fn = os.path.split(pn)
     if ext.lower() not in picture_exts + video_exts:
-        print("{}: ignoring this type of file".format(pn))
         return
     found = found + 1
 
@@ -82,7 +98,7 @@ def archive(pn, recursive):
     else:
         date = None
     if date is None:
-        print("{}: No archive date".format(pn))
+        MyLog("{}: No archive date".format(pn))
         not_archived += 1
         return
 
@@ -91,13 +107,18 @@ def archive(pn, recursive):
     month = months[int(date[1])-1]
     #dbg print("The date in {} is {}".format(fn, ymd))
     folder = dir_archives + '\\' + year + '\\' + month
+
+    # A destination pathname has been computed
+    # if it is the same as the source, we are done.
+    backup_pn = folder + '\\' + fn
+    if pn.lower() == backup_pn.lower():
+        return
     recursive_mkdir(folder)
 
     # Read the CRCs at the destination.
     # If a duplicate CRC exists then delete this file
     crc = crc32(pn)
-    crcs, last_modified = ReadCrcs(folder+"\\crc.csv")
-    backup_pn = folder + '\\' + fn
+    crcs, last_modified = ReadCrcs(log, folder+"\\crc.csv")
     for crcfn in crcs:
         if crc == crcs[crcfn]:
             remove_duplicate(pn, folder+'\\'+crcfn)
@@ -105,8 +126,8 @@ def archive(pn, recursive):
 
     # Try to archive a file simply by renaming it
     if not os.path.exists(backup_pn):
-        os.rename(pn, backup_pn)
         print("{} archived to {}".format(pn, folder))
+        os.rename(pn, backup_pn)
         archived = archived + 1
         return
 
@@ -135,12 +156,12 @@ def archive(pn, recursive):
     not_archived = not_archived + 1
     return
 
-def getfolder(folder, filetype):
+def getfolder(sourcep, filetype):
     """ Pick which folder to use to archive files of a particular file type """
 
     # The returned value must be on the same drive because rename is used
-    folder = os.path.abspath(folder)
-    drive = folder[0:2].upper()
+    sourcep = os.path.abspath(sourcep)
+    drive = sourcep[0:2].upper()
 
     # Start by looking for an environment variable with the same name
     folder = os.environ.get(filetype)
@@ -156,20 +177,25 @@ def getfolder(folder, filetype):
         if folder[:2].upper() == drive:
             return folder + '\\' + filetype
 
-    # If all else fails, use a folder based upon the root with the same name
-    return '\\' + filetype
+    # If all else fails, use a folder based upon the source pathname and filetype
+    rootp, fn = os.path.split(sourcep)
+    return rootp + '\\' + filetype
 
 def main(pn, recursive):
     """ Process files and possibly folders starting with a folder """
+    global found
 
     # Process just the files in a folder
     for fn in glob.glob(glob.escape(pn) + '\\' + filespec):
+        display_update(found)
         if os.path.isfile(fn):
             archive(fn, recursive)
 
     # Process folders recursively, when requested
     if recursive:
         # Note that the saved filespec cannot be used here
+        print("{}: scanning".format(pn))
+        display_update(found, True)
         for pn in glob.glob(glob.escape(pn) + "\\*"):
             if os.path.isdir(pn):
                 main(pn, recursive)
@@ -182,6 +208,9 @@ def help():
     exit()
 
 if __name__ == '__main__':
+    # Create a logfile
+    log = logging("Photo-Backup.log.txt")
+
     # Process the command line arguments
     folder = os.getcwd();   # <path>: use a path other than the current working directory
     recursive = False       # -r: recursively process subfolders
@@ -195,13 +224,14 @@ if __name__ == '__main__':
             folder = arg
         else:
             folder, filespec = os.path.split(arg)
-            folder = os.path.abspath(arg)
+            folder = os.path.abspath(folder)
 
     # Define the alternative locations to archive pictures and movies
     archive_pictures = getfolder(folder, 'Pictures')
     archive_videos = getfolder(folder, 'Videos')
 
     # Find and process folders
+    display_update(found, True)
     main(folder, recursive)
 
     # Print the result
